https://github.com/ImageMagick/CodeScanImageMagickTest/security/code-scanning/9

+++ MagickCore/widget.c
@@ -2036,7 +2036,7 @@ MagickPrivate void XColorBrowserWidget(Display *display,XWindows *windows,
           slider_info.id=0;
         slider_info.y=slider_info.min_y;
         if (colors != 0)
-          slider_info.y+=(int) (slider_info.id*(slider_info.max_y-
+          slider_info.y+=((ssize_t) slider_info.id*(slider_info.max_y-
             slider_info.min_y+1)/colors);
         if (slider_info.id != selection_info.id)
           {
@@ -4646,7 +4646,7 @@ MagickPrivate void XFileBrowserWidget(Display *display,XWindows *windows,
           slider_info.id=0;
         slider_info.y=slider_info.min_y;
         if (files > 0)
-          slider_info.y+=(int) (slider_info.id*(slider_info.max_y-
+          slider_info.y+=((ssize_t) slider_info.id*(slider_info.max_y-
             slider_info.min_y+1)/files);
         if (slider_info.id != selection_info.id)
           {
